+++ router/userRouter.js
@@ -1,46 +1,11 @@
 const express = require('express');
-
-// User Controllers
-const getAllUsers = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined!'
-  })
-}
-
-const createUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined!'
-  })
-}
-
-const getSingleUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined!'
-  })
-}
-
-const updateUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined!'
-  })
-}
-
-const deleteUser = (req, res) => {
-  res.status(500).json({
-    status: 'error',
-    message: 'This route is not yet defined!'
-  })
-}
+const userController = require('../controllers/userController')
 
 // Mounted Routers
-const userRouter = express.Router();
+const router = express.Router();
 
 // User Routes
-userRouter.route('/').get(getAllUsers).post(createUser);
-userRouter.route('/:id').get(getSingleUser).patch(updateUser).delete(deleteUser);
+router.route('/').get(userController.getAllUsers).post(userController.createUser);
+router.route('/:id').get(userController.getSingleUser).patch(userController.updateUser).delete(userController.deleteUser);
 
-module.exports = userRouter;
+module.exports = router;
